Split QuerySessionControl.axaml.cs into partial classes
Move-only refactor; no behavior changes. QuerySessionControl.axaml.cs
(2,251 lines) split into 8 partials:

  Editor       (286) - syntax highlighting, context menu, key/wheel, zoom
  Schema       (355) - ShowSchemaInfoAsync + Format{Indexes,Columns} + BuildWithOptions
  Connection   (120) - Connect_Click + Populate/Fetch* metadata
  Execution    (477) - Execute/EstimatedPlan/GetActualPlan + CaptureAndShowPlan
  Plans        (368) - plan tab add/close/rename/context-menu + Compare picker
  QueryStore   (267) - QueryStore_Click + Overview + OpenQueryStoreForDatabaseAsync
  Advice       ( 24) - Human/Robot advice click + ShowAdviceWindow
  Format       (124) - Format_Click + FormatOptions + CopyRepro

Main file now 177 lines — usings, ctor, fields, AXAML wiring.

Build clean: 0 errors, 0 warnings on PlanViewer.App.
